Move AssetData to package, fix some comments
