+++ LibSaberPatch/AssetDataObjects/AudioClipAssetData.cs
@@ -1,0 +1,96 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
+
+namespace LibSaberPatch.AssetDataObjects
+{
+    public class AudioClipAssetData : AssetData
+    {
+        public const int ClassID = 83;
+
+        public string name;
+        public int loadType;
+        public int channels;
+        public int frequency;
+        public int bitsPerSample;
+        public float length;
+        public bool isTracker;
+
+        public int subsoundIndex;
+        public bool preloadAudio;
+        public bool backgroundLoad;
+        public bool legacy3D;
+
+        public string source;
+        public ulong offset;
+        public ulong size;
+        // 0 = PCM, 1 = Vorbis, 3 = MP3, ...
+        public int compressionFormat;
+
+        public AudioClipAssetData() { }
+
+        public AudioClipAssetData(BinaryReader reader, int _length)
+        {
+            name = reader.ReadAlignedString();
+            loadType = reader.ReadInt32();
+            channels = reader.ReadInt32();
+            frequency = reader.ReadInt32();
+            bitsPerSample = reader.ReadInt32();
+            length = reader.ReadSingle();
+            isTracker = reader.ReadBoolean();
+
+            reader.AlignStream();
+            subsoundIndex = reader.ReadInt32();
+            preloadAudio = reader.ReadBoolean();
+            backgroundLoad = reader.ReadBoolean();
+            legacy3D = reader.ReadBoolean();
+
+            reader.AlignStream();
+            source = reader.ReadAlignedString();
+            offset = reader.ReadUInt64();
+            size = reader.ReadUInt64();
+            compressionFormat = reader.ReadInt32();
+        }
+
+        public override void WriteTo(BinaryWriter w)
+        {
+            w.WriteAlignedString(name);
+            w.Write(loadType);
+            w.Write(channels);
+            w.Write(frequency);
+            w.Write(bitsPerSample);
+            w.Write(length);
+            w.Write(isTracker);
+
+            w.AlignStream();
+            w.Write(subsoundIndex);
+            w.Write(preloadAudio);
+            w.Write(backgroundLoad);
+            w.Write(legacy3D);
+
+            w.AlignStream();
+            w.WriteAlignedString(source);
+            w.Write(offset);
+            w.Write(size);
+            w.Write(compressionFormat);
+        }
+
+        public override int SharedAssetsTypeIndex()
+        {
+            return 5;
+        }
+
+        public override bool Equals(AssetData o)
+        {
+            if (GetType().Equals(o))
+                return source == (o as AudioClipAssetData).source && name == (o as AudioClipAssetData).name;
+            return false;
+        }
+
+        public override List<string> OwnedFiles(SerializedAssets assets)
+        {
+            return new List<string>() { source };
+        }
+    }
+}
