+++ LibSaberPatch/AssetDataObjects/GameObjectAssetData.cs
@@ -1,0 +1,56 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
+
+namespace LibSaberPatch.AssetDataObjects
+{
+    public class GameObjectAssetData : AssetData
+    {
+        public const int ClassID = 1;
+
+        public AssetPtr[] components;
+        public uint layer;
+        public string name;
+        public ushort tag;
+        public bool isActive;
+
+        public GameObjectAssetData(BinaryReader reader, int _length)
+        {
+            int size = reader.ReadInt32();
+            components = new AssetPtr[size];
+            for (int i = 0; i < size; i++)
+            {
+                components[i] = new AssetPtr(reader);
+            }
+            layer = reader.ReadUInt32();
+            name = reader.ReadAlignedString();
+            tag = reader.ReadUInt16();
+            isActive = reader.ReadBoolean();
+        }
+
+        public override bool Equals(AssetData o)
+        {
+            //TODO Implement
+            return false;
+        }
+
+        public override int SharedAssetsTypeIndex()
+        {
+            return 0;
+        }
+
+        public override void WriteTo(BinaryWriter w)
+        {
+            w.Write(components.Length);
+            foreach (AssetPtr p in components)
+            {
+                p.WriteTo(w);
+            }
+            w.Write(layer);
+            w.WriteAlignedString(name);
+            w.Write(tag);
+            w.Write(isActive);
+        }
+    }
+}
