+++ LibSaberPatch/Utils.cs
@@ -9,6 +9,7 @@ using SixLabors.ImageSharp.PixelFormats;
 using LibSaberPatch.BehaviorDataObjects;
 using System.Collections.Generic;
 using System.Linq;
+using LibSaberPatch.AssetDataObjects;
 
 namespace LibSaberPatch
 {
