+++ tests/SerializedAssetTests.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.IO.Compression;
 using System.Reflection;
 using Newtonsoft.Json;
+using LibSaberPatch.AssetDataObjects;
 
 namespace TestApp
 {
